airport logic update

+++ src/main/java/com/dmytrobozhor/airlinereservationservice/service/AbstractAirportService.java
@@ -5,7 +5,7 @@ import com.dmytrobozhor.airlinereservationservice.domain.FlightDetail;
 
 import java.util.Optional;
 
-public interface AbstractAirportService extends AbstractCrudService<Airport, Integer> {
+public interface AbstractAirportService extends AbstractCrudService<Airport, Long> {
 
     Optional<Airport> findByAllFields(Airport airport);
